requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-SocketIO==5.4.1
fonttools==4.54.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==3.3.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
simple-websocket==1.1.0
six==1.16.0
Werkzeug==3.1.1
wsproto==1.2.0
